Minor fixes (avoiding compilation warning with double -> int)
Ignore-this: 8e2d0d3c18475df78a1a569f16db6d31

darcs-hash:20111110203325-e973b-6ed44addce52e1fb0bd4c21d7e350fe5e4487bf8.gz

+++ CutManager/src/MuonSelection.cc
@@ -104,7 +104,7 @@ void MuonSelection::LockCuts(){
 		}
 		else if( cut->first == "MinNValidHitsSATrk" )
 		{
-			kMinNValidHitsSATrk = cut->second;
+			kMinNValidHitsSATrk = (int)cut->second;
 		}
 		else if( cut->first == "MaxNormChi2GTrk" )
 		{
@@ -112,15 +112,15 @@ void MuonSelection::LockCuts(){
 		}
 		else if( cut->first == "MinNumOfMatches" )
 		{
-			kMinNumOfMatches = cut->second;
+			kMinNumOfMatches = (int)cut->second;
 		}
 		else if( cut->first == "MinNValidPixelHitsInTrk" )
 		{
-			kMinNValidPixelHitsInTrk = cut->second;
+			kMinNValidPixelHitsInTrk = (int)cut->second;
 		}
 		else if( cut->first == "MinNValidHitsInTrk" )
 		{
-			kMinNValidHitsInTrk = cut->second;
+			kMinNValidHitsInTrk = (int)cut->second;
 		}
 		else if( cut->first == "MaxDeltaPtMuOverPtMu" )
 		{
